Rename vitess package as it hosts docker test infrastructure for other data stores
