+++ misk-hibernate/src/main/kotlin/misk/database/StartVitessService.kt
@@ -1,4 +1,4 @@
-package misk.vitess
+package misk.database
 
 import misk.jdbc.DataSourceConfig
 import kotlin.reflect.KClass
